Update 07._Animations.md
Repaired a link for Animatable.  I used a search url... but I don't know if that is wise.  If we target a specific version of Animatable (ie. 2.3), then this link goes stale when 2.4 (and beyond) is released.  It would be nice to have a... umm... virtual path? for the link.  Something like...

http://doc.babylonjs.com/classes/latest/Animatable

'latest' would point-to the latest version of the classes.  Anyway, I have never done an edit where I didn't use my own GitHub account to do the edit-with.  This is a...

"You’re editing a file in a project you don’t have write access to. We’ve created a fork of this project for you to commit your proposed changes to. Submitting a change to this file will write it to a new branch in your fork, so you can send a pull request."

...I don't know if that is wise, or not.  Noob!  :)  I've never had a "commit discussion" with GitHub, either, so if someone would like to talk with me... in one of those, I'd be willing to try it.  :)  Thanks.