Added API links to Observables.md
Hi guys! Because this is my first contribution to the documentation please check whether I included the links correctly. This change belong to the issue #1770 . I think it would be too much to place links at every playe where an API function is mentioned. So I only linked the first occurence like in wikipedia. Are the changes in that way useful or do I miss something?