Update Use_HDR_Environment.md
Added a different code sample for rotating environments which is visible in the inspector and reworded the section about dropping .hdr files into the sandbox. Also moved the rotation text to the first section of general information about adding an environment instead of being slotted in the section about the external prefiltering tools.